Update 1_💰_Commission_Calculator.py
(1) we should add your new security / authentication model / code(2) on the results summary, i want to have a "total amount due" that shows up. and also a total amount due adjusted amount that  ignores the values less than zero and add up only values owed to salespeople that are greater than zero. (3) on the summary tab of the excel sheet, i'd like to add a column for total amount due as well. it should ignore amounts less than $0 and sum up the amounts due that are higher than zero. i've put a little picture (4) on the tab for each salesperson i'd like to have the by category info above the ledger. i have included a potential layout for that. (5) we should add the ability to email the report like we have done for the other apps.